Another nudge towards Pocket PC support
git-svn-id: https://appconference.svn.sourceforge.net/svnroot/appconference/trunk@494 f854266a-e20d-0410-a9be-8852184c940f

+++ iaxclient/lib/gsm/inc/private.h
@@ -152,7 +152,7 @@ static __inline__ short GSM_SUB(short a, short b)
 /* Nonportable, but faster: */
 
 #define	GSM_ADD(a, b)	\
-	((ulongword)((ltmp = (longword)(a) + (longword)(b)) - MIN_WORD) > \
+	(short) ((ulongword)((ltmp = (longword)(a) + (longword)(b)) - MIN_WORD) > \
 		MAX_WORD - MIN_WORD ? (ltmp > 0 ? MAX_WORD : MIN_WORD) : ltmp)
 
 # define GSM_SUB(a, b)	\
